cloudfunction/requirements.txt
google-cloud-vision==0.42.0
google-cloud-storage==1.26.0
google-cloud-pubsub==2.13.6
Wand==0.5.9